refactor: 불필요한 공백 제거, import 제거, 메소드명 변경
MovingStatus의 move 메서드명을 behave로 변경
