+++ src/main/java/racingcar/model/Distance.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public class Distance {
     private int distance = 0;
