+++ src/main/java/racingcar/model/MovingRule.java
@@ -1,10 +1,9 @@
-package racingcar;
+package racingcar.model;
 
 public class MovingRule {
-
     public static final int MOVING_STANDARD_NUMBER = 4;
 
-    public MovingStatus move(int i) {
+    public MovingStatus behave(int i) {
         if ( i < MOVING_STANDARD_NUMBER) {
             return MovingStatus.STOP;
         }
