+++ src/main/java/racingcar/model/MovingStatus.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public enum MovingStatus {
     GO, STOP;
