+++ src/main/java/racingcar/model/Race.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import racingcar.util.StringUtils;
