+++ src/main/java/racingcar/model/RacingCars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Set;
 
 public class RacingCars {
-
     private static final String CAR_DUP_ERROR_MESSAGE = "[ERROR] 자동차 이름은 중복될 수 없습니다.";
 
     private List<Car> racingCars;
