+++ src/main/java/racingcar/model/Round.java
@@ -1,9 +1,8 @@
-package racingcar;
+package racingcar.model;
 
 import racingcar.util.ValidationUtils;
 
 public class Round {
-
     private static final String ROUND_GO_ERROR_MESSAGE = "[ERROR] 남은 시도 횟수는 0입니다.";
     private static final String ROUND_EMPTY_ERROR_MESSAGE = "[ERROR] 시도 횟수는 필수값입니다.";
     private static final String ROUND_RANGE_ERROR_MESSAGE = "[ERROR] 시도 횟수는 1-2147483647 사이 숫자만 가능합니다.";
