+++ src/main/java/racingcar/util/StringUtils.java
@@ -4,7 +4,6 @@ import java.util.Arrays;
 import java.util.List;
 
 public class StringUtils {
-
     public static List<String> splitByDelimiter(String text, String delimiter) {
         text = text.replaceAll(" ", "");
         String[] splitResultList = text.split(delimiter, -1);
