+++ src/main/java/racingcar/util/ValidationUtils.java
@@ -4,7 +4,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class ValidationUtils {
-
     private static final Pattern alphaPattern = Pattern.compile("[a-zA-Z]+");
     private static final Pattern numericPattern = Pattern.compile("^[-]?[0-9]+");
